requirements.txt
pytest  # test framework
pytest-cov  # coverage plugin
hypothesis  # property-based testing
requests  # HTTP client
PyYAML  # YAML parsing
pytest-socket  # restrict network calls in tests
responses  # mock HTTP responses
rich  # colorized CLI output
typer  # CLI framework
click  # dependency for typer
jsonschema>=3.2  # JSON validation
pydantic>=2  # data models
fastapi  # web framework
httpx  # async HTTP client
matplotlib  # graphs for docs
seaborn  # enhanced plotting
numpy  # numerical computing
pydeps  # dependency graphs
sphinx  # documentation generator
sphinx_rtd_theme  # docs theme
uvicorn  # ASGI server
pytest-env  # env var plugin
aiohttp>=3.9.0  # async HTTP engine
structlog  # structured logging
sentry-sdk  # error reporting
bandit[toml]  # security scanner
